[FIX] account_chart_update_multilang: proper translating of HTML fields
The algorithm to translate them is done by terms, which is a bit clunky for our use case.

However, since `account.fiscal.position.template`'s note field is Text, not Html, and that's the one we care about, we can predict that the Html version will just wrap it with a <p> tag.

Well, in short, it's ugly but works.

MT-7597