[FIX] base_vat_optional_vies: don't block mass checks
The whole point of this module is to allow VIES failures. However, when running mass checks, if one single partner failed, the whole process had to start again.

FWIW that can be also problematic due to rate limits in VIES service. For that same reason, we now skip partners that are validated already, and those that are will inherit the status from their parent anyway.

Now, the process will just continue and report failures at the end.

It includes new tests, removes dead code, and reduces pre-existing tests verbosity.

@moduon MT-7763